Include schema directives to ensure proper schema validation
