+++ src/Execution/Arguments/ArgumentSetFactory.php
@@ -10,7 +10,7 @@ use GraphQL\Type\Definition\ResolveInfo;
 use InvalidArgumentException;
 use Nuwave\Lighthouse\Schema\AST\ASTBuilder;
 use Nuwave\Lighthouse\Schema\AST\ASTHelper;
-use Nuwave\Lighthouse\Schema\Factories\DirectiveFactory;
+use Nuwave\Lighthouse\Schema\DirectiveLocator;
 
 class ArgumentSetFactory
 {
@@ -25,14 +25,14 @@ class ArgumentSetFactory
     protected $argumentTypeNodeConverter;
 
     /**
-     * @var \Nuwave\Lighthouse\Schema\Factories\DirectiveFactory
+     * @var \Nuwave\Lighthouse\Schema\DirectiveLocator
      */
     protected $directiveFactory;
 
     public function __construct(
         ASTBuilder $astBuilder,
         ArgumentTypeNodeConverter $argumentTypeNodeConverter,
-        DirectiveFactory $directiveFactory
+        DirectiveLocator $directiveFactory
     ) {
         $this->documentAST = $astBuilder->documentAST();
         $this->argumentTypeNodeConverter = $argumentTypeNodeConverter;
