+++ src/Schema/Factories/DirectiveFactory.php
@@ -2,217 +2,62 @@
 
 namespace Nuwave\Lighthouse\Schema\Factories;
 
-use GraphQL\Language\AST\DirectiveNode;
-use GraphQL\Language\AST\Node;
-use GraphQL\Language\Parser;
-use Illuminate\Support\Arr;
-use Illuminate\Support\Collection;
-use Illuminate\Support\Str;
-use Nuwave\Lighthouse\Exceptions\DirectiveException;
-use Nuwave\Lighthouse\Schema\DirectiveNamespacer;
-use Nuwave\Lighthouse\Schema\Directives\BaseDirective;
-use Nuwave\Lighthouse\Support\Contracts\Directive;
-use Nuwave\Lighthouse\Support\Utils;
+use GraphQL\Language\AST\DirectiveDefinitionNode;
+use GraphQL\Type\Definition\Directive;
+use GraphQL\Type\Definition\FieldArgument;
+use Nuwave\Lighthouse\Schema\AST\ASTHelper;
+use Nuwave\Lighthouse\Schema\ExecutableTypeNodeConverter;
 
 class DirectiveFactory
 {
     /**
-     * A map from short directive names to full class names.
-     *
-     * E.g.
-     * [
-     *   'create' => 'Nuwave\Lighthouse\Schema\Directives\CreateDirective',
-     *   'custom' => 'App\GraphQL\Directives\CustomDirective',
-     * ]
-     *
-     * @var string[]
+     * @var \Nuwave\Lighthouse\Schema\ExecutableTypeNodeConverter
      */
-    protected $resolvedClassnames = [];
+    protected $definitionNodeConverter;
 
-    /**
-     * The paths used for locating directive classes.
-     *
-     * Should be tried in the order they are contained in this array,
-     * going from the most significant to least significant.
-     *
-     * @var string[]
-     */
-    protected $directiveNamespaces;
-
-    /**
-     * @var \Nuwave\Lighthouse\Schema\DirectiveNamespacer
-     */
-    protected $directiveNamespacer;
-
-    public function __construct(DirectiveNamespacer $directiveNamespacer)
+    public function __construct(ExecutableTypeNodeConverter $definitionNodeConverter)
     {
-        $this->directiveNamespacer = $directiveNamespacer;
+        $this->definitionNodeConverter = $definitionNodeConverter;
     }
 
     /**
-     * Create a directive by the given directive name.
+     * Transform node to type.
      */
-    public function create(string $directiveName): Directive
+    public function handle(DirectiveDefinitionNode $directive): Directive
     {
-        $directiveClass = $this->resolve($directiveName);
+        $arguments = [];
+        /** @var \GraphQL\Language\AST\InputValueDefinitionNode $argument */
+        foreach ($directive->arguments as $argument) {
+            /** @var \GraphQL\Type\Definition\Type&\GraphQL\Type\Definition\InputType $argumentType */
+            $argumentType = $this->definitionNodeConverter->convert($argument->type);
 
-        return app($directiveClass);
-    }
+            $fieldArgumentConfig = [
+                'name' => $argument->name->value,
+                'description' => data_get($argument->description, 'value'),
+                'type' => $argumentType,
+            ];
 
-    /**
-     * Resolve the class for a given directive name.
-     *
-     *
-     * @throws \Nuwave\Lighthouse\Exceptions\DirectiveException
-     */
-    protected function resolve(string $directiveName): string
-    {
-        if ($directiveClass = Arr::get($this->resolvedClassnames, $directiveName)) {
-            return $directiveClass;
-        }
-
-        if (! $this->directiveNamespaces) {
-            $this->directiveNamespaces = $this->directiveNamespacer->gather();
-        }
-
-        foreach ($this->directiveNamespaces as $baseNamespace) {
-            $directiveClass = $baseNamespace.'\\'.static::className($directiveName);
-
-            if (class_exists($directiveClass)) {
-                if (! is_a($directiveClass, Directive::class, true)) {
-                    throw new DirectiveException("Class $directiveClass must implement the interface ".Directive::class);
-                }
-
-                $this->addResolved($directiveName, $directiveClass);
-
-                return $directiveClass;
+            if ($defaultValue = $argument->defaultValue) {
+                $fieldArgumentConfig += [
+                    'defaultValue' => ASTHelper::defaultValueForArgument($defaultValue, $argumentType),
+                ];
             }
+
+            $arguments [] = new FieldArgument($fieldArgumentConfig);
         }
 
-        throw new DirectiveException("No directive found for `{$directiveName}`");
-    }
-
-    /**
-     * Returns the expected class name for a directive name.
-     */
-    protected static function className(string $directiveName): string
-    {
-        return Str::studly($directiveName).'Directive';
-    }
-
-    /**
-     * Returns the expected directive name for a class name.
-     */
-    public static function directiveName(string $className): string
-    {
-        $baseName = basename(str_replace('\\', '/', $className));
-
-        return lcfirst(
-            Str::before($baseName, 'Directive')
-        );
-    }
-
-    /**
-     * @deprecated use the RegisterDirectiveNamespaces event instead, this method will be removed as of v5
-     * @see \Nuwave\Lighthouse\Events\RegisterDirectiveNamespaces
-     *
-     * @return $this
-     */
-    public function addResolved(string $directiveName, string $className): self
-    {
-        // Bail to respect the priority of namespaces, the first
-        // resolved directive is kept
-        if (in_array($directiveName, $this->resolvedClassnames, true)) {
-            return $this;
+        $locations = [];
+        // Might be a NodeList, so we can not use array_map()
+        foreach ($directive->locations as $location) {
+            $locations[] = $location->value;
         }
 
-        $this->resolvedClassnames[$directiveName] = $className;
-
-        return $this;
-    }
-
-    /**
-     * @return $this
-     */
-    public function setResolved(string $directiveName, string $className): self
-    {
-        $this->resolvedClassnames[$directiveName] = $className;
-
-        return $this;
-    }
-
-    /**
-     * @deprecated will be removed as of v5
-     * @return $this
-     */
-    public function clearResolved(): self
-    {
-        $this->resolvedClassnames = [];
-
-        return $this;
-    }
-
-    /**
-     * Get all directives of a certain type that are associated with an AST node.
-     *
-     * @return \Illuminate\Support\Collection<\Nuwave\Lighthouse\Support\Contracts\Directive> of type <$directiveClass>
-     */
-    public function createAssociatedDirectivesOfType(Node $node, string $directiveClass): Collection
-    {
-        return $this
-            ->createAssociatedDirectives($node)
-            ->filter(Utils::instanceofMatcher($directiveClass));
-    }
-
-    /**
-     * Get all directives that are associated with an AST node.
-     *
-     * @return \Illuminate\Support\Collection<\Nuwave\Lighthouse\Support\Contracts\Directive>
-     */
-    public function createAssociatedDirectives(Node $node): Collection
-    {
-        return (new Collection($node->directives))
-            ->map(function (DirectiveNode $directiveNode) use ($node): Directive {
-                $directive = $this->create($directiveNode->name->value);
-
-                if ($directive instanceof BaseDirective) {
-                    $directive->hydrate($directiveNode, $node);
-                }
-
-                return $directive;
-            });
-    }
-
-    /**
-     * Get a single directive of a type that belongs to an AST node.
-     *
-     * Use this for directives types that can only occur once, such as field resolvers.
-     * This throws if more than one such directive is found.
-     *
-     * TODO rename to exclusiveDirective
-     *
-     * @throws \Nuwave\Lighthouse\Exceptions\DirectiveException
-     */
-    public function createSingleDirectiveOfType(Node $node, string $directiveClass): ?Directive
-    {
-        $directives = $this->createAssociatedDirectivesOfType($node, $directiveClass);
-
-        if ($directives->count() > 1) {
-            $directiveNames = $directives
-                ->map(function (Directive $directive): string {
-                    $definition = Parser::directiveDefinition(
-                        $directive::definition()
-                    );
-
-                    return '@'.$definition->name->value;
-                })
-                ->implode(', ');
-
-            throw new DirectiveException(
-                "Node {$node->name->value} can only have one directive of type {$directiveClass} but found [{$directiveNames}]."
-            );
-        }
-
-        return $directives->first();
+        return new Directive([
+            'name' => $directive->name->value,
+            'description' => data_get($directive->description, 'value'),
+            'locations' => $locations,
+            'args' => $arguments,
+            'astNode' => $directive,
+        ]);
     }
 }
