+++ src/Validation/RulesDirective.php
@@ -28,7 +28,7 @@ directive @rules(
   Specified as an input object that maps rules to messages,
   e.g. { email: "Must be a valid email", max: "The input was too long" }
   """
-  messages: [RulesMessageMap!]
+  messages: [String!]
 ) on ARGUMENT_DEFINITION | INPUT_FIELD_DEFINITION
 SDL;
     }
