+++ src/Validation/RulesForArrayDirective.php
@@ -26,7 +26,7 @@ directive @rulesForArray(
   Specified as an input object that maps rules to messages,
   e.g. { email: "Must be a valid email", max: "The input was too long" }
   """
-  messages: RulesMessageMap
+  messages: [String!]
 ) on ARGUMENT_DEFINITION | INPUT_FIELD_DEFINITION
 SDL;
     }
